requirements-local.txt
click==6.7
coverage==4.4.1
Flask==0.12.2
Flask-Restless==0.17.0
flask-restless-swagger-2==0.0.3
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mimerender==0.6.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2

